src/backend/requirements.txt
# Auth0 Enterprise Platform - FastAPI Backend
# Python 3.11+

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async Support
anyio>=4.2.0
httpx>=0.26.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Caching
redis>=5.0.0

# Background Tasks
celery>=5.3.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0

# HTTP Client with Retry
tenacity>=8.2.0

# Utilities
python-dateutil>=2.8.0
orjson>=3.9.0

# Rate Limiting
slowapi>=0.1.9

# CORS
starlette>=0.35.0
